+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.4.0 / 2018-07-15
+
+* Add tournaments endpoint
+
+## 0.3.4 / 2018-07-14
+
+* Add blue zone custom options telemetry object
+
+* Add missing attributes from V1.2.0, V1.3.0, V1.4.0
+
+* Remove deprecated data from V2.0.0
+
+## 0.3.3 / 2018-07-11
+
+* Remove _V key from V2.0.0 - @wookienz
+
+* Add implementation of new telemetry events - @Tetraquark
+
 ## 0.3.1 / 2018-06-09
 
 * Add timeout to requests - @vgustov